refactor: modularize analysis helpers
