+++ analysis/analysis_pipeline.py
@@ -39,6 +39,11 @@ from .syntheticETFBuilder import (
     combine_surfaces,
     build_synthetic_iv as build_synthetic_iv_pillars,
 )
+from .synthetic import build_synthetic_iv_series
+from .relative_value import (
+    relative_value_atm_report_corrweighted,
+    latest_relative_snapshot_corrweighted,
+)
 
 from .beta_builder.beta_builder import (
     pca_weights,
@@ -203,15 +208,6 @@ def build_synthetic_surface(
     surfaces = build_surfaces(tickers=list(w.keys()), cfg=cfg, most_recent_only=most_recent_only)
     return combine_surfaces(surfaces, w)
 
-
-def build_synthetic_iv_series(
-    weights: Mapping[str, float],
-    pillar_days: Union[int, Iterable[int]] = DEFAULT_PILLARS_DAYS,
-    tolerance_days: float = 7.0,
-) -> pd.DataFrame:
-    """Create a weighted ATM pillar IV time series."""
-    w = {k.upper(): float(v) for k, v in weights.items()}
-    return build_synthetic_iv_pillars(w, pillar_days=pillar_days, tolerance_days=tolerance_days)
 
 # -----------------------------------------------------------------------------
 # Weights (modern unified first, legacy fallback kept)
@@ -455,120 +451,7 @@ def save_betas(
         return [p]
     return save_correlations(mode=mode, benchmark=benchmark, base_path=base_path)
 # =========================
-# Relative value (target vs synthetic peers by corr)
-# -----------------------------------------------------------------------------
-def _fetch_target_atm(
-    target: str,
-    pillar_days: Iterable[int],
-    tolerance_days: float = 7.0
-) -> pd.DataFrame:
-    atm = load_atm()
-    atm = atm[atm["ticker"] == target].copy()
-    if atm.empty:
-        return pd.DataFrame(columns=["asof_date", "pillar_days", "iv"])
-    piv = nearest_pillars(atm, pillars_days=list(pillar_days), tolerance_days=tolerance_days)
-    out = piv.groupby(["asof_date", "pillar_days"])["iv"].mean().rename("iv").reset_index()
-    return out[["asof_date", "pillar_days", "iv"]]
-
-
-def _rv_metrics_join(target_iv: pd.DataFrame, synth_iv: pd.DataFrame, lookback: int = 60) -> pd.DataFrame:
-    tgt = target_iv.rename(columns={"iv": "iv_target"})
-    syn = synth_iv.rename(columns={"iv": "iv_synth"})
-    df = pd.merge(tgt, syn, on=["asof_date", "pillar_days"], how="inner").sort_values(["pillar_days", "asof_date"])
-    if df.empty:
-        return df
-
-    def per_pillar(g: pd.DataFrame) -> pd.DataFrame:
-        g = g.copy()
-        g["spread"] = g["iv_target"] - g["iv_synth"]
-        roll = max(5, int(lookback // 5))
-        m = g["spread"].rolling(lookback, min_periods=roll).mean()
-        s = g["spread"].rolling(lookback, min_periods=roll).std(ddof=1)
-        g["z"] = (g["spread"] - m) / s
-        # percentile rank of latest value within window
-        def _pct_rank(x: pd.Series) -> float:
-            return x.rank(pct=True).iloc[-1]
-        g["pct_rank"] = g["spread"].rolling(lookback, min_periods=roll).apply(_pct_rank, raw=False)
-        return g
-
-    return df.groupby("pillar_days", group_keys=False).apply(per_pillar, include_groups=False)
-
-
-def relative_value_atm_report_corrweighted(
-    target: str,
-    peers: Iterable[str] | None = None,
-    mode: str = "iv_atm",                # 'ul' | 'iv_atm' | 'surface'  (weights source)
-    pillar_days: Iterable[int] = DEFAULT_PILLARS_DAYS,
-    lookback: int = 60,
-    tolerance_days: float = 7.0,
-    weight_power: float = 1.0,
-    clip_negative: bool = True,
-) -> tuple[pd.DataFrame, pd.Series]:
-    """
-    Compute relative value (spread/z/pct) using correlation-based peer weights.
-    Returns (rv_dataframe, weights_used).
-    """
-    w = peer_weights_from_correlations(
-        benchmark=target,
-        peers=peers,
-        mode=mode,
-        pillar_days=pillar_days,
-        tenor_days=DEFAULT_TENORS,
-        mny_bins=DEFAULT_MNY_BINS,
-        clip_negative=clip_negative,
-        power=weight_power,
-    )
-    if w.empty:
-        empty_cols = ["asof_date", "pillar_days", "iv_target", "iv_synth", "spread", "z", "pct_rank"]
-        return pd.DataFrame(columns=empty_cols), w
-
-    synth = build_synthetic_iv_series(weights=w.to_dict(), pillar_days=pillar_days, tolerance_days=tolerance_days)
-    tgt = _fetch_target_atm(target, pillar_days=pillar_days, tolerance_days=tolerance_days)
-    rv = _rv_metrics_join(tgt, synth, lookback=lookback)
-    return rv, w
-
-
-def latest_relative_snapshot_corrweighted(
-    target: str,
-    peers: Iterable[str] | None = None,
-    mode: str = "iv_atm",
-    pillar_days: Iterable[int] = (7, 30, 60, 90),
-    lookback: int = 60,
-    tolerance_days: float = 7.0,
-    **kwargs,
-) -> tuple[pd.DataFrame, pd.Series]:
-    """
-    Convenience: last date per pillar with RV metrics and the weights used.
-    """
-    rv, w = relative_value_atm_report_corrweighted(
-        target=target, peers=peers, mode=mode, pillar_days=pillar_days,
-        lookback=lookback, tolerance_days=tolerance_days, **kwargs
-    )
-    # If nothing came back, just return early
-    if rv.empty:
-        return rv, w
-
-    # Ensure required columns exist even if upstream fallback omitted them
-    # - pillar_days: create a placeholder if missing (single bucket)
-    # - asof_date: ensure it exists and is sortable
-    if "pillar_days" not in rv.columns:
-        # treat whole snapshot as a single synthetic pillar bucket (e.g., 0)
-        rv = rv.copy()
-        rv["pillar_days"] = 0
-    if "asof_date" not in rv.columns:
-        # cannot manufacture real dates; create a constant so sort/group don't explode
-        rv = rv.copy()
-        rv["asof_date"] = pd.Timestamp("1970-01-01")
-
-    # Keep only columns we know how to present; tolerate missing metric columns
-    base_cols = ["asof_date", "pillar_days", "iv_target", "iv_synth", "spread", "z", "pct_rank"]
-    present_cols = [c for c in base_cols if c in rv.columns]
-
-    # Group safely by pillar and take the latest per pillar
-    last = rv.sort_values("asof_date").groupby("pillar_days", dropna=False).tail(1)
-    # Guarantee pillar_days in the projected set
-    proj_cols = sorted(set(["pillar_days", "asof_date"] + present_cols))
-    return last[proj_cols].sort_values("pillar_days"), w
+# Relative value helpers reside in ``analysis.relative_value``
 
 # -----------------------------------------------------------------------------
 # Basic DB helpers + caches
