+++ analysis/beta_builder/utils.py
@@ -1,4 +1,5 @@
 import numpy as np
+import pandas as pd
 
 
 def impute_col_median(X: np.ndarray) -> np.ndarray:
@@ -17,3 +18,18 @@ def zscore_cols(X: np.ndarray) -> tuple[np.ndarray, np.ndarray, np.ndarray]:
     sd = np.nanstd(X, axis=0, ddof=1, keepdims=True)
     sd = np.where(~np.isfinite(sd) | (sd <= 0), 1.0, sd)
     return (X - mu) / sd, mu, sd
+
+
+def safe_var(x: pd.Series) -> float:
+    """Finite variance or ``nan`` if unavailable/degenerate."""
+    v = x.var()
+    return float(v) if (v is not None and np.isfinite(v) and v > 0) else float("nan")
+
+
+def beta(df: pd.DataFrame, x: str, b: str) -> float:
+    """Simple ``beta`` of column ``x`` versus benchmark column ``b``."""
+    a = df[[x, b]].dropna()
+    if len(a) < 5:
+        return float("nan")
+    vb = safe_var(a[b])
+    return float(a[x].cov(a[b]) / vb) if np.isfinite(vb) else float("nan")
